Work on test example for API. Not quite working with supertest
