+++ api/package.json
@@ -4,11 +4,10 @@
   "description": "",
   "main": "dist/server.js",
   "scripts": {
-    "test": "node ./node_modules/tape/bin/tape tests/**; exit 0;",
+    "test": "./node_modules/.bin/babel-tape-runner tests/**",
     "lint": "node ./node_modules/eslint/bin/eslint src/**; exit 0;",
     "QA": "npm run lint && npm run test",
     "start": "npm run build && node dist/app",
-    "lint": "node ./node_modules/eslint/bin/eslint src/**; exit 0;",
     "dev": "./node_modules/.bin/nodemon --exec ./node_modules/.bin/babel-node --debug --presets es2015 --presets stage-1 -- src/app.js",
     "build": "./node_modules/.bin/babel src -d dist"
   },
@@ -19,7 +18,7 @@
     "koa-json": "^1.1.1",
     "koa-logger": "^1.3.0",
     "koa-mount": "^1.3.0",
-    "koa-route": "^2.4.2",
+    "koa-parse-json": "^1.0.1",
     "koa-router": "^5.3.0",
     "node-pg-migrate": "0.0.9",
     "pg-promise": "^2.8.5"
@@ -31,9 +30,11 @@
     "babel-polyfill": "^6.3.14",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-stage-1": "^6.3.13",
+    "babel-tape-runner": "^2.0.0",
     "eslint": "^1.10.3",
     "eslint-config-google": "^0.3.0",
     "nodemon": "^1.8.1",
+    "superagent": "^1.6.1",
     "supertest": "^1.1.0",
     "tape": "^4.4.0"
   }
