+++ api/src/app.js
@@ -1,3 +1,18 @@
-require('babel-polyfill');
+import koa from 'koa'
+import logger from 'koa-logger'
+import json from 'koa-json'
+import parseJson from 'koa-parse-json'
 
-require('./server');
+import cranes from './routers/cranes'
+
+const app = koa();
+
+app.use(parseJson());
+app.use(json());
+app.use(logger());
+
+app
+.use(cranes.routes())
+  .use(cranes.allowedMethods());
+
+export { app }
