+++ api/src/server.js
@@ -1,20 +1,7 @@
-import koa from 'koa'
-import logger from 'koa-logger'
-import json from 'koa-json'
-import parseJson from 'koa-parse-json'
-
-import cranes from './routers/cranes'
+import { app } from './app'
 
 const PORT = process.env.PORT || 3000;
 
-const app = koa();
+const server = app.listen(PORT);
 
-app.use(parseJson());
-app.use(json());
-app.use(logger());
-
-app
-  .use(cranes.routes())
-  .use(cranes.allowedMethods());
-
-app.listen(PORT);
+export { server }
